MODELSIM-5 Add integration tests
