+++ .github/workflows/full-ci-cd.yml
@@ -1,0 +1,124 @@
+name: Full CI/CD
+on:
+  push:
+  pull_request:
+  workflow_dispatch:
+    inputs:
+      release_version:
+        description: 'Version to release (i.e.: 1.2.0)'
+        required: true
+      next_version:
+        description: 'Next development cycle version (i.e.: 1.3.0-SNAPSHOT)'
+        required: true
+
+jobs:
+  build:
+    name: Build on Commit
+    runs-on: ${{ matrix.os }}
+    strategy:
+      matrix:
+        os: [ ubuntu-latest ]
+        sonarqube: [ 7.9.4, 8.4.2.36762 ]
+
+    steps:
+      - name: Git Checkout
+        uses: actions/checkout@v2
+        with:
+          fetch-depth: 0
+
+      - name: Set up JDK 11
+        uses: actions/setup-java@v1
+        with:
+          java-version: 11
+
+      - name: Cache Maven Dependencies
+        uses: actions/cache@v2
+        with:
+          path: ~/.m2
+          key: ${{ runner.os }}-m2-${{ hashFiles('**/pom.xml') }}
+          restore-keys: ${{ runner.os }}-m2
+
+      - name: Set Release Version ${{ github.event.inputs.release_version }}
+        if: github.event_name == 'workflow_dispatch' && matrix.sonarqube == '7.9.4'
+        run: mvn -B versions:set -DnewVersion=${{ github.event.inputs.release_version }} -DgenerateBackupPoms=false
+
+      - name: Build and Test against SonarQube ${{ matrix.sonarqube }}
+        run: mvn -B clean verify -Dsonar.runtimeVersion=${{ matrix.sonarqube }}
+
+#      - name: Run SonarQube Analysis
+#        if: contains(github.ref, 'master') && matrix.sonarqube == '7.9.4'
+#        run: mvn -B sonar:sonar -Dsonar.login=xxx
+
+      - name: Upload Plugin JAR
+        if: matrix.sonarqube == '7.9.4'
+        uses: actions/upload-artifact@v2
+        with:
+          name: plugin
+          path: |
+            sonar-coverage-modelsim-plugin/target/sonar-coverage-modelsim-plugin-*.jar
+            !sonar-coverage-modelsim-plugin/target/sonar-coverage-modelsim-plugin-*-sources.jar
+
+      - name: Commit, Push and Tag Release Version ${{ github.event.inputs.release_version }}
+        if: github.event_name == 'workflow_dispatch' && matrix.sonarqube == '7.9.4'
+        uses: EndBug/add-and-commit@v5
+        with:
+          add: '*'
+          author_name: GitHub Actions
+          author_email: github-actions@linty-services.com
+          message: Release version ${{ github.event.inputs.release_version }}
+          tag: ${{ github.event.inputs.release_version }}
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Set Next Development Cycle Version ${{ github.event.inputs.next_version }}
+        if: github.event_name == 'workflow_dispatch' && matrix.sonarqube == '7.9.4'
+        run: mvn -B versions:set -DnewVersion=${{ github.event.inputs.next_version }} -DgenerateBackupPoms=false
+
+      - name: Commit and Push Next Development Cycle Version ${{ github.event.inputs.next_version }}
+        if: github.event_name == 'workflow_dispatch' && matrix.sonarqube == '7.9.4'
+        uses: EndBug/add-and-commit@v5
+        with:
+          add: '*'
+          author_name: GitHub Actions
+          author_email: github-actions@linty-services.com
+          message: Bump version to ${{ github.event.inputs.next_version }}
+        env:
+          GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+  pre-release:
+    name: GitHub Pre-release
+    runs-on: ubuntu-latest
+    needs: build
+    if: contains(github.ref, 'master')
+    steps:
+      - name: Download Plugin JAR
+        uses: actions/download-artifact@v2
+        with:
+          name: plugin
+      - name: "GitHub Pre-release: latest"
+        uses: marvinpinto/action-automatic-releases@latest
+        with:
+          repo_token: "${{ secrets.GITHUB_TOKEN }}"
+          automatic_release_tag: "latest"
+          prerelease: true
+          title: Latest build from master
+          files: sonar-coverage-modelsim-plugin-*.jar
+
+  release:
+    name: GitHub Release
+    runs-on: ubuntu-latest
+    needs: build
+    if: github.event_name == 'workflow_dispatch'
+    steps:
+      - name: Download Plugin JAR
+        uses: actions/download-artifact@v2
+        with:
+          name: plugin
+      - name: "GitHub Release: ${{ github.event.inputs.release_version }}"
+        uses: marvinpinto/action-automatic-releases@latest
+        with:
+          repo_token: "${{ secrets.GITHUB_TOKEN }}"
+          automatic_release_tag: ${{ github.event.inputs.release_version }}
+          prerelease: false
+          title: ${{ github.event.inputs.release_version }}
+          files: sonar-coverage-modelsim-plugin-*.jar
