+++ modelsim-its/plugin/src/test/java/com/lintyservices/sonar/modelsim/its/Tests.java
@@ -1,0 +1,88 @@
+/*
+ * SonarQube Linty ModelSim :: Integration Tests :: Plugin
+ * Copyright (C) 2019-2020 Linty Services
+ * mailto:contact@linty-services.com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+ */
+package com.lintyservices.sonar.modelsim.its;
+
+import com.sonar.orchestrator.Orchestrator;
+import com.sonar.orchestrator.build.SonarScanner;
+import com.sonar.orchestrator.http.HttpMethod;
+import com.sonar.orchestrator.locator.FileLocation;
+import org.junit.ClassRule;
+import org.junit.runner.RunWith;
+import org.junit.runners.Suite;
+import org.sonarqube.ws.WsMeasures;
+import org.sonarqube.ws.client.HttpConnector;
+import org.sonarqube.ws.client.WsClient;
+import org.sonarqube.ws.client.WsClientFactories;
+import org.sonarqube.ws.client.measure.ComponentWsRequest;
+
+import java.io.File;
+import java.util.List;
+
+import static java.util.Collections.singletonList;
+
+@RunWith(Suite.class)
+@Suite.SuiteClasses({
+  MetricsTest.class,
+})
+public class Tests {
+
+  @ClassRule
+  public static final Orchestrator ORCHESTRATOR = Orchestrator.builderEnv()
+    .setSonarVersion(System.getProperty("sonar.runtimeVersion", "DEV"))
+    .addPlugin(FileLocation.byWildcardMavenFilename(new File("../../sonar-coverage-modelsim-plugin/target"), "sonar-coverage-modelsim-plugin-*.jar"))
+    .build();
+
+  public static SonarScanner createSonarScannerBuild() {
+    return SonarScanner.create()
+      .setProjectVersion("1.0")
+      .setSourceEncoding("UTF-8")
+      .setSourceDirs("src");
+  }
+
+  public static void setProfile(String profileName, String projectKey) {
+    ORCHESTRATOR.getServer().associateProjectToQualityProfile(projectKey, "vhdl", profileName);
+  }
+
+  public static void resetData() {
+    ORCHESTRATOR.getServer().newHttpCall("api/orchestrator/reset")
+      .setMethod(HttpMethod.POST)
+      .setAdminCredentials()
+      .execute();
+  }
+
+  static WsMeasures.Measure getMeasure(String componentKey, String metricKey) {
+    WsMeasures.ComponentWsResponse response = newWsClient().measures().component(new ComponentWsRequest()
+      .setComponent(componentKey)
+      .setMetricKeys(singletonList(metricKey)));
+    List<WsMeasures.Measure> measures = response.getComponent().getMeasuresList();
+    return measures.size() == 1 ? measures.get(0) : null;
+  }
+
+  static Double getMeasureAsDouble(String componentKey, String metricKey) {
+    WsMeasures.Measure measure = getMeasure(componentKey, metricKey);
+    return (measure == null) ? null : Double.parseDouble(measure.getValue());
+  }
+
+  static WsClient newWsClient() {
+    return WsClientFactories.getDefault().newClient(HttpConnector.newBuilder()
+      .url(ORCHESTRATOR.getServer().getUrl())
+      .build());
+  }
+}
