+++ modelsim-its/plugin/src/test/resources/com/lintyservices/sonar/vhdl/its/profiles/empty-profile.xml
@@ -1,0 +1,6 @@
+<?xml version='1.0' encoding='UTF-8'?>
+<profile>
+  <name>empty-profile</name>
+  <language>vhdl</language>
+  <rules/>
+</profile>
