+++ sonar-coverage-modelsim-plugin/pom.xml
@@ -1,0 +1,147 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <groupId>com.lintyservices.sonar-coverage-modelsim-plugin</groupId>
+    <artifactId>parent</artifactId>
+    <version>1.4.0-SNAPSHOT</version>
+  </parent>
+
+  <artifactId>sonar-coverage-modelsim-plugin</artifactId>
+  <packaging>sonar-plugin</packaging>
+  <name>SonarQube Linty ModelSim :: Plugin</name>
+  <url>https://www.linty-services.com</url>
+
+  <properties>
+    <license.owner>Linty Services</license.owner>
+    <license.mailto>mailto:contact@linty-services.com</license.mailto>
+    <sonar.version>6.7</sonar.version>
+    <jacoco.version>0.8.4</jacoco.version>
+    <sonar.host.url>https://sonar.linty-services.com</sonar.host.url>
+  </properties>
+
+  <organization>
+    <name>Linty Services</name>
+    <url>https://www.linty-services.com</url>
+  </organization>
+
+  <inceptionYear>2019</inceptionYear>
+  <licenses>
+    <license>
+      <name>GNU LGPL 3</name>
+      <url>http://www.gnu.org/licenses/lgpl.txt</url>
+    </license>
+  </licenses>
+
+  <developers>
+    <developer>
+      <id>racodond</id>
+      <name>David Racodon</name>
+      <email>david.racodon@linty-services.com</email>
+      <organization>Linty Services</organization>
+      <organizationUrl>https://www.linty-services.com</organizationUrl>
+    </developer>
+    <developer>
+      <id>louisv</id>
+      <name>Vincent Louis</name>
+      <email>vincent.louis@linty-services.com</email>
+      <organization>Linty Services</organization>
+      <organizationUrl>https://www.linty-services.com</organizationUrl>
+    </developer>
+    <developer>
+      <id>rocam</id>
+      <name>Mathias Roca</name>
+      <email>mathias.roca@linty-services.com</email>
+      <organization>Linty Services</organization>
+      <organizationUrl>https://www.linty-services.com</organizationUrl>
+    </developer>
+  </developers>
+
+  <scm>
+    <connection>scm:git:github.com:Linty-Services/sonar-coverage-modelsim-plugin.git</connection>
+    <developerConnection>scm:git:github.com:Linty-Services/sonar-coverage-modelsim-plugin.git</developerConnection>
+    <url>https://github.com/Linty-Services/sonar-coverage-modelsim-plugin.git</url>
+  </scm>
+
+  <issueManagement>
+    <system>jira</system>
+    <url>https://s2ceh-vhdl.atlassian.net/browse/MODELSIM</url>
+  </issueManagement>
+
+  <dependencies>
+    <dependency>
+      <groupId>org.sonarsource.sonarqube</groupId>
+      <artifactId>sonar-plugin-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.apache.maven</groupId>
+      <artifactId>maven-project</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>com.google.guava</groupId>
+      <artifactId>guava</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>commons-io</groupId>
+      <artifactId>commons-io</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>commons-lang</groupId>
+      <artifactId>commons-lang</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
+    </dependency>
+    <dependency>
+      <groupId>org.sonarsource.sonarqube</groupId>
+      <artifactId>sonar-testing-harness</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.easytesting</groupId>
+      <artifactId>fest-assert</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-all</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-core</artifactId>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>ch.qos.logback</groupId>
+      <artifactId>logback-classic</artifactId>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
+
+  <build>
+    <testResources>
+      <testResource>
+        <directory>sonar-coverage-modelsim-plugin/src/main/resources</directory>
+      </testResource>
+      <testResource>
+        <directory>src/test/resources</directory>
+      </testResource>
+    </testResources>
+
+    <plugins>
+      <plugin>
+        <groupId>org.sonarsource.sonar-packaging-maven-plugin</groupId>
+        <artifactId>sonar-packaging-maven-plugin</artifactId>
+        <configuration>
+          <pluginName>ModelSim</pluginName>
+          <pluginClass>com.lintyservices.sonar.plugins.modelsim.ModelsimPlugin</pluginClass>
+          <sonarQubeMinVersion>${sonar.version}</sonarQubeMinVersion>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
+</project>
