+++ sonar-coverage-modelsim-plugin/src/test/resources/com/lintyservices/sonar/plugins/modelsim/ModelsimSensorTest/commons-chain-coverage.xml
