+++ sonar-coverage-modelsim-plugin/src/test/resources/com/lintyservices/sonar/plugins/modelsim/ModelsimSensorTest/wrong-coverage.xml
